pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "xattrs"
version = "0.0.1"
description = 'An eXtensible serializing and deserializing toolkit for Python strucuted classes'
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
keywords = ["attrs", "cattrs", "serde", "json", "schema"]
authors = [{ name = "Lanqing Huang", email = "lqhuang@outlook.com" }]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX",
  "Typing :: Typed",
]
dependencies = ["attrs", "typing_extensions", "datetype"]

[project.optional-dependencies]
yaml = ["ruamel.yaml>0.18"]
toml = ["tomlkit"]
ujson = ["ujson"]
msgpack = ["msgspec"]
all = ["xattrs[toml]", "xattrs[yaml]", "xattrs[ujson]", "xattrs[msgpack]"]
test = [
  "hypothesis",
  "pympler",
  "pytest",
  "pytest-cov",
  'mypy',
  'pytest-mypy-plugins',
  "coverage[toml]",
]
docs = [
  "towncrier",
  "furo",
  "myst-parser",
  "sphinx",
  "sphinx-notfound-page",
  "sphinx-copybutton",
  "sphinx-autobuild",
  "sphinxcontrib-towncrier",
  "sphinx-autodoc-typehints",
  "sphinxext-opengraph",
]
linter = ['mypy', 'black', 'isort', 'ruff', "pyupgrade", "pre-commit"]
build = ["pip", "build>=1.0.0", "wheel", "pdm", "pdm-backend"]
dev = ["xattrs[linter,test,build,docs]"]


[project.urls]
Documentation = "https://github.com/lqhuang/xattrs#readme"
Issues = "https://github.com/lqhuang/xattrs/issues"
Source = "https://github.com/lqhuang/xattrs"

# ## Build
[tool.hatch.build.targets.wheel]
packages = ["src/xattrs"]

[tool.pdm.build]
package-dir = "src"
source-includes = [
  "docs",
  "src",
  "tests",
  "CHANGELOG.md",
  "LICENSE",
  "Makefile",
  "pyproject.toml",
  "README.md",
]

## Pytest
[tool.pytest.ini_options]
addopts = ["-ra", "--strict-markers", "--strict-config"]
xfail_strict = true
testpaths = "tests"

[tool.check-wheel-contents]
toplevel = ["xattrs"]

## Formatter
[tool.black]
line-length = 88
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
known_typing = [
  "types",
  "typing",
  "typing_extensions",
  "collections.abc",
  "xattrs._compat.typing",
  "xattrs._typing",
  "xattrs.typing",
]
sections = [
  "FUTURE",
  "TYPING",
  "STDLIB",
  "THIRDPARTY",
  "FIRSTPARTY",
  "LOCALFOLDER",
]

## Linter
[tool.mypy]
## doc: https://mypy.readthedocs.io/en/stable/config_file.html
strict = true

## Untyped definitions and calls
check_untyped_defs = true
## Miscellaneous
warn_unused_configs = true
## Configuring error messages
show_column_numbers = true
show_error_context = true
show_error_codes = true
# pretty = true  # disable for IDE parser; try to add cli option in Makefile
## Incremental mode
cache_fine_grained = true
## Advanced options
warn_incomplete_stub = true

disable_error_code = ['empty-body', 'no-untyped-def']

[[tool.mypy.overrides]]
module = ["xattrs._compat.*", "xattrs.typing"]
disable_error_code = ['empty-body']

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
disable_error_code = ["no-untyped-def", "arg-type"]

[[tool.mypy.overrides]]
module = "examples.*"
disallow_untyped_defs = false

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md#main-pyright-config-options
typeCheckingMode = "basic"
reportGeneralTypeIssues = false
# reportPrivateUsage = false

[tool.ruff]
line-length = 88
indent-width = 4
fix = false

typing-modules = ["xattrs._compat.typing"]

# [tool.ruff.format]
# docstring-code-format = true
# docstring-code-line-length = "dynamic"
# line-ending = "lf"
# quote-style = "double"

[tool.ruff.lint]
# Enable rules:
# Pyflakes (F), pycodestyle (E, W), pydocstyle (D), pyupgrade (UP),
# flake8-bandit (S), flake8-bugbear (B), flake8-comprehensions (C4), flake8-datetimez (DTZ), flake8-pytest-style (PT)
# Pylint (PLC, PLE, PLR, PLW), Ruff-specific rules (RUF)
select = ["F", "E", "W", "UP", "S", "B", "C4", "DTZ", "PT"]
extend-select = ["PLC", "PLE", "PLR", "PLW", "RUF"]

ignore = [
  "E501",    # Line too long (95 > 88 characters)
  "S101",    # Use of `assert` detected
  "RUF100",  # Unused `noqa` directive for: E305 (not implemented)
  "PLR2004", # Magic value used in comparison, consider replacing with a constant variable
  "PLR0913", # Too many statements
  "C408",
  "B027",    # Allow non-abstract empty methods in abstract base classes
  "PLC0414", # Import alias does not rename original package
]


#  A list of check code prefixes to consider autofix-able.
fixable = ["F", "E711", "UP", "B"]
unfixable = [
  "I",
  "F401", # Don't touch unused imports
]

ignore-init-module-imports = true

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"] # imported but unused
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]

[tool.ruff.lint.pydocstyle]
convention = "pep257"
